refactor(connect): move management method definitions to corresponding folder
