--- packages/connect-common/src/types/api/getSettings.ts
@@ -1,4 +1,0 @@
-import type { Response } from '../params';
-import type { ConnectSettings } from '../settings';
-
-export declare function getSettings(): Response<ConnectSettings>;
