+++ packages/connect-common/src/types/api/management/getFirmwareHash.ts
@@ -1,6 +1,6 @@
 import type { MessagesSchema as PROTO } from '@trezor/protobuf';
 
-import type { Params, Response } from '../params';
+import type { Params, Response } from '../../params';
 
 export declare function getFirmwareHash(
     params: Params<PROTO.GetFirmwareHash>,
