+++ packages/connect-common/src/types/api/management/getNonce.ts
@@ -1,5 +1,5 @@
 import type { Nonce } from '@trezor/protobuf';
 
-import type { CommonParams, Response } from '../params';
+import type { CommonParams, Response } from '../../params';
 
 export declare function getNonce(params?: CommonParams): Response<Nonce>;
