+++ packages/connect-common/src/types/api/management/getSettings.ts
@@ -1,0 +1,4 @@
+import type { Response } from '../../params';
+import type { ConnectSettings } from '../../settings';
+
+export declare function getSettings(): Response<ConnectSettings>;
