+++ packages/connect-common/src/types/api/management/pingDevice.ts
@@ -4,6 +4,6 @@
 
 import type { MessagesSchema as PROTO } from '@trezor/protobuf';
 
-import type { Params, Response } from '../params';
+import type { Params, Response } from '../../params';
 
 export declare function pingDevice(params: Params<PROTO.Ping>): Response<PROTO.Ping>;
