+++ packages/connect-common/src/types/api/management/recoveryDevice.ts
@@ -6,7 +6,7 @@ import { MessagesSchema as PROTO } from '@trezor/protobuf';
 import type { Static } from '@trezor/schema-utils';
 import { Type } from '@trezor/schema-utils';
 
-import type { Params, Response } from '../params';
+import type { Params, Response } from '../../params';
 
 export type RecoveryDevice = Static<typeof RecoveryDevice>;
 export const RecoveryDevice = Type.Composite([
