+++ packages/connect-common/src/types/api/management/setBusy.ts
@@ -4,6 +4,6 @@
 
 import type { MessagesSchema as PROTO } from '@trezor/protobuf';
 
-import type { Params, Response } from '../params';
+import type { Params, Response } from '../../params';
 
 export declare function setBusy(params: Params<PROTO.SetBusy>): Response<PROTO.Success>;
