+++ packages/connect-common/src/types/api/management/telemetryGet.ts
@@ -1,5 +1,5 @@
 import type { MessagesSchema as PROTO } from '@trezor/protobuf';
 
-import type { Params, Response } from '../params';
+import type { Params, Response } from '../../params';
 
 export declare function telemetryGet(params: Params<PROTO.TelemetryGet>): Response<PROTO.Telemetry>;
