+++ packages/connect-common/src/types/api/management/thpGetCredentials.ts
@@ -1,5 +1,5 @@
 import type { ThpCredentials } from '@trezor/protocol';
 
-import type { CommonParams, Response } from '../params';
+import type { CommonParams, Response } from '../../params';
 
 export declare function thpGetCredentials(params?: CommonParams): Response<ThpCredentials>;
