+++ packages/connect-common/src/types/api/management/thpRemoveCredentials.ts
@@ -1,7 +1,7 @@
 import type { MessagesSchema as PROTO } from '@trezor/protobuf';
 import type { ThpCredentials } from '@trezor/protocol';
 
-import type { CommonParams, Response } from '../params';
+import type { CommonParams, Response } from '../../params';
 
 export declare function thpRemoveCredentials(
     params: CommonParams & { credentials?: ThpCredentials[] },
