+++ packages/connect-common/src/types/api/management/wipeDevice.ts
@@ -4,6 +4,6 @@
 
 import type { MessagesSchema as PROTO } from '@trezor/protobuf';
 
-import type { CommonParams, Response } from '../params';
+import type { CommonParams, Response } from '../../params';
 
 export declare function wipeDevice(params?: CommonParams): Response<PROTO.Success>;
